Changed the range of image pixel values from [-1, 1] to [0, 1]
This change was made for two reasons:

1. The Frechet Inception Distance object requires the pixel values to be in range [0, 1].
By having image pixel values lie in the range [0, 1], we avoid having to unnormalize the
values before computing the FID score. We also avoid having to unnormalize images before plotting
images using matplotlibs imshow().

2. This allows us to use a neat little trick to improve performance. By not normalizing the pixel values
we can store the images in the dataloader as datatype uint8 instead of float32. This speeds up the transfer
of images from cpu to gpu in the training loop. Once the images have been transferred, we can then simply
scale the images by 1 / 255.0.