Fixed bugs where torchinfo summary and FrechetInceptionDistance could cause some tensors to be on different device than DEVICE
1. If no device argument is provided to torchinfo summary, it would default to using cuda:0 if available,
which would cause issues if we wanted to use cpu instead.

2. If FrechetInceptionDistance would load an existing FID object that was updated on a different device than
the one we're currently using it would cause some attributes to be on the wrong device. Hence, we move the FID
object to DEVICE after having initialised and processed it to make sure all attributes are on the same device.